+++ lib/src/networks/multilayer_perceptron.dart
@@ -5,8 +5,5 @@ class MultilayerPerceptron extends Network {
   MultilayerPerceptron({
     required List<int> layerSizes,
     required ActivationAlgorithm activationAlgorithm,
-  }) : super(
-    layerSizes: layerSizes,
-    activationAlgorithm: activationAlgorithm
-  );
+  }) : super(layerSizes: layerSizes, activationAlgorithm: activationAlgorithm);
 }