+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 # Version 0.1.0
 
 - Added Multi-layer Perceptron and a basic algorithm for backpropagation
+
+# Version 0.1.1
+
+- Added README.md, updated formatting
+
+# Version 0.2.0
+
+- Added FF ( feedforward ) and simple Perceptron networks
+- Added LReLU, eLU and tanh activation function
+- Renamed 'sigmoid' to 'logistic' function